Fix #38 Allowing silient and private activities
This removes the specification that activities without an audience
should be federated to the default audience for the user. This now adds
a note which provides a suggested behavour in the two cases of where no
audience is specified or where just the public audience is
specified. This is a note as we want to leave it up to the
implementation to do what works best for them but provide guidance in
cases where this isn't important.